Fix subsetting range bug
Fixes a RuntimeError when using the --subset option with
a CSV that has too few species. This error only occurred
when k > species. This change will return less than k entries
when there are too few species. This matches the behavior
of heapq.nlargest() used for non species grouping.

Fixes #87